Fix five cross-language defects found by external audit
- Node charset: Go/Python/sh only rejected whitespace/hyphens, so nodes like
  "node.01" validated in three implementations and failed in the other three.
  All six now enforce the spec charset [A-Za-z0-9_]+; the EBNF is relaxed to
  allow a leading underscore, which every implementation already accepted.
  Two adversarial node fixtures added (43 conformance cases per impl).
- Rust: ParsedWid.sequence and ParsedHlcWid.logical_counter were u32, so a
  spec-legal all-nines W=10 sequence failed to parse only in Rust. Now i64.
- TypeScript: Math.pow(10,18)-1 is not representable (evaluates to 10^18) and
  sequence increments stall above 2^53; maxSeq/maxLC are capped at
  MAX_SAFE_INTEGER-1 so rollover stays sound, and the parsed sequence field
  documents its precision limit (validation is string-exact and unaffected).
- Go: crypto/rand failure silently fell back to time-derived padding,
  defeating the collision defense -- now fatal; sequence parse uses checked
  64-bit ParseInt instead of an ignored Atoi; parsed fields are int64.
- SECURITY.md claimed only the sh implementation exposes the w-otp secret via
  process arguments; every implementation does for inline KEY=. Documented,
  along with the KEY path/secret ambiguity, recommending the file form.

Verified: all six agree on the node probes; Rust parses W=18 all-nines
exactly; full per-language batteries, id-conformance 43/43 x6, wotp-parity,
smoke_crypto 6x6, and stream conformance all pass.

+++ rust/src/hlc.rs
@@ -15,7 +15,9 @@ use crate::wid::{TimeUnit, WidError};
 pub struct ParsedHlcWid {
     pub raw: String,
     pub timestamp: DateTime<Utc>,
-    pub logical_counter: u32,
+    /// i64, not u32: the spec allows W up to 18, and an all-nines W=10
+    /// counter (9999999999) already exceeds u32::MAX.
+    pub logical_counter: i64,
     pub node: String,
     pub padding: Option<String>,
 }
@@ -116,7 +118,7 @@ pub fn parse_hlc_wid_with_unit(
     }
 
     let timestamp = parse_ts(time_unit, date_str, time_str).ok_or(WidError::InvalidTimestamp)?;
-    let logical_counter: u32 = lc_str
+    let logical_counter: i64 = lc_str
         .parse()
         .map_err(|_| WidError::InvalidFormat(wid.to_string()))?;
